common: Gate various types behind feature flags
